// frontend/src/components/types.d.ts
export class EventClass {
    Category: string = "";
    Date_time: Date = new Date();
    Downvote: number = 0;
    EventDescription: string = "";
    EventStatus: string = "";
    Event_time: Date = new Date();
    ID: number = 0;
    Severity: string = "";
    Title: string = "";
    Upvote: number = 0;
    UserID: number = 0;
    UserName: string = "";
    isVerified: number = 0;
}